app: service-commons: extract and generalize dual-write transaction logic
Extracting the fixed logic from the CustomerService, moving it into a parent
module and generalizing it into a reusable function allows us to make sure
every CRUD operation + event emission on any entity will be handled in
the same way.

The new control flow is:

```
Controller.createEntity(RequestBody body) {
  body.flatMap(entity -> service.createEntity(entity, eventFactory, messageBroker);
}
->
Service.createEntity(entity, eventFactory, messageBroker) {
  return DualWriteTransactionHelper.createEntityControlFlowTemplate(
    template,
    transactionalOperator,
    entity,
    log,
    eventFactory,
    messageBroker,
    queryFunction
}
->
DualWriteTransactionHelper.createEntityControlFlowTemplate() {
  template.insert(entity)
    .doFirst(logPendingWriteToDB)
    .flatMap(readEntityFromDB)
    .doOnSuccess(logCompletedTransactionToDB)
    .delayUntil(beforeTransactionFinalizeTryEventEmitCallback)
}
```

Reading the entity from DB is needed (with R2DBC) to ensure that no
error has happened during the INSERT statement. But also, to get the
customer ID into the event.
With this control flow, an error during the DB write will not emit the
event to the broker - therefore avoiding that wrong facts are escaping -
but also an error during the event emission will rollback the DB
transaction - ensuring, that the dual-write transaction can only
"create" data if everything can be executed successfully.

Closes: #11