app: business-intelligence-service: improve message <-> json conversion mechanism
Event types will now be auto-deduced by a newer Jackson version.
See `@JsonInfoType`.
This removes the need to write and register a new MessageConverter
inside the business-intelligence-service for each new event type and has
the added benefit that each subsequent new service that also needs to do
event processing does not have to declare these MessageConverters.

Sadly, `use = Id.DEDUCTION` did not work, as our events are flatten in
the event object -> json step and each event has it's payload written
under the key 'data'. Jackson should still be able to deduce which data
json value corresponds to which event object, but it's a fairly new
tech, so bugs are expected.

Closes: #12