app: build: checkstyle: raise violationSeverity to error
This avoids an bug due to an unconfigurable formatting plugin which, to
my surprise, seems to do a "wrong" formatting decision, which in turn
trips up the checkstyle-plugin as it expects the code to be formatted in
a different way.

Specifically this occurs for the following code:

@JsonSubTypes(
    @Type(...)
    @Type(...)
)

As seen above, the formatting tool indents each new nested annotation
with 4 spaces, but the checkstyle plugin expects it to have only 2
spaces.