app: event-processing: switch from JSON to Apache Avro event data representation
The JSON data representation had the issue that we still needed JSON
Schema to allow multiple event types on the same topic.
Because JSON Schema is rather painful to work with, we are switching to
Apache Avro.

In this data representation, data types are defined through *.avsc files
which in turn auto-generate Java classes through the avro-compiler.
These classes handle anything from en/decoding, how the data can be
accessed and how they need to be de/serialized.

This allows us to do the following:

* Define schemas for each event type in a clear and readable matter,
* Define schemas for each Kafka topic to allow multiple event types to
be written to the same topic.
* Not care how to serialize/deserialize, because this handled by the
Avro compiler.

This is an important step, as allowing multiple event types on the same
topic is a necessity when using Event Sourcing as we want all possible
events for a specific entity to be on the same Kafka topic, as this
makes sure that they are ordered in the correct way.
Explicitly this means, that a CustomerCreatedEvent cannot occur before
the CustomerDeletedEvent from the perspective of the consumer, as they
have to be ordered in the same way they have been added to the topic by
the producer.
Using different Kafka topics for each event type would complicate this.