app: customer-service: [PoC] implement update customer functionality
This is kind of a weird idea, but using Javers, it is possible to
generate diffs between entities, which would be really nice to decide,
which fields exactly have changed and only omit those changed in the
CustomerUpdatedEvent (which is not yet implemented).

This is simply a PoC and WiP to check, if the idea is at all possible.
Which would be nice. Afterwards, it would be possible to allow an client
to only sent parts of an customer, i.e. only the fields he would like to
update.