app: events: implement builder pattern like interface
This allows us to supply an entity before the event is created, which
comes in handy when we want to create update events, as currently, the
updated entity only holds fields that were actually updated and
therefore differs from the entity that will be inserted as that one
actually holds all values.
This allows us to instruct our dualWriteTransactionHelper to use an
other entity during the event creation instead of the one read from the
DB, without introducing too much change to the current logic.