app: business-intelligence-service: allow multiple of this service to consume events.
Being in the same consumer groups forces each service to compete for the
same events, i.e. only one application in the group can process this
event.
This is something good, if you have multiple services that might write
something back down to a shared DB or something.
But here, not needed.