app: service-commons: refactor from calling setters via reflection to setting fields via reflection
More reliable, and works with enum classes.
Enum classes in Java are a bit special as they are all anonymous, static
classes.
Therefore, they have a specific subclass, i.e. OrderState$1 instead of
OrderState.
That makes it errornous to pass them to setters that would accept
OrderState, as the reflection API wants to find an method for
OrderState$1.